feat(api): Refactor authentication and profile management APIs
- Implemented new API client utilities for token management and attachment normalization.
- Added contact API service for sending messages.
- Created jobs API service for job listings, applications, and saving jobs.
- Developed profile API service for user profile CRUD operations and profile picture management.
- Introduced settings API service for password changes, account management, and user settings updates.
- Added employer and job seeker dashboard components with dynamic imports for better optimization.
- Updated existing authentication functions to handle errors and responses more effectively.
- Removed deprecated attachment types and unnecessary site configuration function.
