+++ app/dashboard/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 import React from "react";
 import { useUserProfile } from "@/hooks/useUserProfile";
-import JobSeekerDashboard from "./job_seeker";
-import EmployerDashboard from "./employer";
+import JobSeekerDashboard from "@/components/pages/dashboard/JobSeekerDashboard";
+import EmployerDashboard from "@/components/pages/dashboard/EmployerDashboard";
 import { useRouter } from "next/navigation";
 import { DashboardSkeleton } from "@/components/ui/Skeleton";
 
@@ -18,9 +18,11 @@ const Dashboard = () => {
     return (
       <div className="flex flex-col min-h-screen bg-gray-50">
         <div className="flex flex-col items-center justify-center flex-1 p-8 text-center">
-          <h2 className="text-2xl font-bold text-gray-900 mb-2">Unable to load profile</h2>
+          <h2 className="text-2xl font-bold text-gray-900 mb-2">
+            Unable to load profile
+          </h2>
           <p className="text-gray-500 mb-6">Please try logging in again.</p>
-          <button 
+          <button
             onClick={() => router.push("/signin")}
             className="px-6 py-2 bg-primary text-white rounded-lg hover:bg-primary-hover transition-colors"
           >
