--- components/pages/auth/index.ts
@@ -1,9 +1,0 @@
-export { FormField } from "@/components/ui/FormField";
-export { PasswordField } from "@/components/ui/PasswordField";
-export { RoleCard } from "@/components/ui/RoleCard";
-export { FileUploadZone } from "@/components/ui/FileUploadZone";
-export { AuthHeader } from "@/components/ui/AuthHeader";
-export { AuthFooter } from "@/components/ui/AuthFooter";
-export { StepIndicator } from "./signup/StepIndicator";
-
-export { SignupWizard } from "./signup/SignupWizard";
