+++ components/pages/dashboard/EmployerDashboard.tsx
@@ -1,6 +1,5 @@
 "use client";
 import React from "react";
-import PageHeader from "@/components/ui/PageHeader";
 
 const EmployerDashboard = () => {
   return (
